Fixed core dump from stoi errors

+++ Project10/FinalProject.cpp
@@ -6,6 +6,7 @@
 
 #include <iostream>
 #include <fstream>
+#include <ctype.h>
 #include "BinaryTree.h"
 
 using namespace std;
@@ -27,7 +28,7 @@ void functionsMenu(BinaryTree& tree)
 {
     string inputLine;
     int secondChoice = 0;
-    
+
     while (secondChoice != 4)
     {
         cout << endl << "=====Tree Functions=====" << endl;
@@ -35,36 +36,66 @@ void functionsMenu(BinaryTree& tree)
         cout << "2. Lookup a value" << endl;
         cout << "3. Delete a value" << endl;
         cout << "4. Main Menu" << endl;
-        
+
         getline(cin, inputLine);
-        secondChoice = stoi(inputLine);
-        
-        switch (secondChoice)
+
+        if (isdigit(inputLine[0]))
         {
-            case 1:
-                cout << "Insert value:" << endl;
-                getline(cin, inputLine);
-                tree.insert(stoi(inputLine));
-                break;
-            case 2:
+            secondChoice = stoi(inputLine);
+
+            switch (secondChoice)
             {
-                cout << "Lookup value:" << endl;
-                getline(cin, inputLine);
-                tree.lookup(stoi(inputLine));
-                TreeNode* node = tree.lookup(stoi(inputLine));
-                if (node == NULL)
-                    cout << "Value not found." << endl;
-                else
-                    cout << "Value found! " << node->key << endl;
+                case 1:
+                    cout << "Insert value:" << endl;
+                    getline(cin, inputLine);
+
+                    if (isdigit(inputLine[0]))
+                    {
+                        tree.insert(stoi(inputLine));
+                    }
+                    else
+                    {
+                        cout << "Please input an integer" << endl;
+                    }
+                    break;
+                case 2:
+                {
+                    cout << "Lookup value:" << endl;
+                    getline(cin, inputLine);
+                    if (isdigit(inputLine[0]))
+                    {
+                    tree.lookup(stoi(inputLine));
+                    TreeNode* node = tree.lookup(stoi(inputLine));
+                    if (node == NULL)
+                        cout << "Value not found." << endl;
+                    else
+                        cout << "Value found! " << node->key << endl;
+                    }
+                    else
+                    {
+                        cout << "Please input an integer" << endl;
+                    }
+                    break;
+                }
+                case 3:
+                    cout << "Delete value:" << endl;
+                    getline(cin, inputLine);
+                    if (isdigit(inputLine[0]))
+                    {
+                        tree.deleteNode(stoi(inputLine));
+                    }
+                    else
+                    {
+                        cout << "Please input an integer" << endl;
+                    }
+                    break;
+                case 4:
+                    break;
             }
-                break;
-            case 3:
-                cout << "Delete value:" << endl;
-                getline(cin, inputLine);
-                tree.deleteNode(stoi(inputLine));
-                break;
-            case 4:
-                break;
+        }
+        else
+        {
+            cout << "Please input an integer" << endl;
         }
     }
 }
@@ -77,7 +108,7 @@ void informationMenu(BinaryTree& tree)
 
     while (secondChoice != 6)
     {
-    
+
         cout << endl <<"=====Tree Information=====" << endl;
         cout << "1. Tree Size" << endl;
         cout << "2. Max Depth " << endl;
@@ -85,32 +116,39 @@ void informationMenu(BinaryTree& tree)
         cout << "4. Max Value" << endl;
         cout << "5. Valid Tree" << endl;
         cout << "6. Main Menu" << endl;
-        
+
         getline(cin, inputLine);
-        secondChoice = stoi(inputLine);
-        
-        switch (secondChoice)
+        if (isdigit(inputLine[0]))
         {
-            case 1:
-                cout << "Tree size: " << tree.size() << endl;
-                break;
-            case 2:
-                cout << "Max depth: " << tree.maxDepth() << endl;
-                break;
-            case 3:
-                cout << "Min value: " << tree.minValue() << endl;
-                break;
-            case 4:
-                cout << "Max value: " << tree.maxValue() << endl;
-                break;
-            case 5:
-                if (tree.isBST())
-                    cout << "The tree is valid" << endl;
-                else
-                    cout << "The tree is not valid." << endl;
-                break;
-            case 6:
-                break;
+            secondChoice = stoi(inputLine);
+
+            switch (secondChoice)
+            {
+                case 1:
+                    cout << "Tree size: " << tree.size() << endl;
+                    break;
+                case 2:
+                    cout << "Max depth: " << tree.maxDepth() << endl;
+                    break;
+                case 3:
+                    cout << "Min value: " << tree.minValue() << endl;
+                    break;
+                case 4:
+                    cout << "Max value: " << tree.maxValue() << endl;
+                    break;
+                case 5:
+                    if (tree.isBST())
+                        cout << "The tree is valid" << endl;
+                    else
+                        cout << "The tree is not valid." << endl;
+                    break;
+                case 6:
+                    break;
+            }
+        }
+        else
+        {
+            cout << "Please input an integer" << endl;
         }
     }
 }
@@ -120,7 +158,7 @@ void operationsMenu(BinaryTree& tree)
 {
     string inputLine;
     int secondChoice = 0;
-    
+
     while (secondChoice != 4)
     {
         cout << endl << "=====Tree Operations=====" << endl;
@@ -128,32 +166,46 @@ void operationsMenu(BinaryTree& tree)
         cout << "2. Mirror Tree" << endl;
         cout << "3. Find Path Sum" << endl;
         cout << "4. Main Menu" << endl;
-        
+
         getline(cin, inputLine);
-        secondChoice = stoi(inputLine);
-        
-        switch (secondChoice)
+        if (isdigit(inputLine[0]))
         {
-            case 1:
-                tree.doubleTree();
-                cout << "Doubled Tree: " << endl;
-                tree.prettyPrint();
-                break;
-            case 2:
-                tree.mirror();
-                cout << "Mirrored Tree: " << endl;
-                tree.prettyPrint();
-                break;
-            case 3:
-                cout << "Enter value:" << endl;
-                getline(cin, inputLine);
-                if (tree.hasPathSum(stoi(inputLine)))
-                    cout << "The answer is yes." << endl;
-                else
-                    cout << "The answer is no." << endl;
-                break;
-            case 4:
-                break;
+            secondChoice = stoi(inputLine);
+
+            switch (secondChoice)
+            {
+                case 1:
+                    tree.doubleTree();
+                    cout << "Doubled Tree: " << endl;
+                    tree.prettyPrint();
+                    break;
+                case 2:
+                    tree.mirror();
+                    cout << "Mirrored Tree: " << endl;
+                    tree.prettyPrint();
+                    break;
+                case 3:
+                    cout << "Enter value:" << endl;
+                    getline(cin, inputLine);
+                    if (isdigit(inputLine[0]))
+                    {
+                        if (tree.hasPathSum(stoi(inputLine)))
+                            cout << "The answer is yes." << endl;
+                        else
+                            cout << "The answer is no." << endl;
+                    }
+                    else
+                    {
+                        cout << "Please input an integer" << endl;
+                    }
+                    break;
+                case 4:
+                    break;
+            }
+        }
+        else
+        {
+            cout << "Please input an integer" << endl;
         }
     }
 }
@@ -163,7 +215,7 @@ void printingMenu(BinaryTree& tree)
 {
     string inputLine;
     int secondChoice = 0;
-    
+
     while (secondChoice != 5)
     {
         cout << endl << "=====Print the Tree=====" << endl;
@@ -172,26 +224,33 @@ void printingMenu(BinaryTree& tree)
         cout << "3. In Order Traversal" << endl;
         cout << "4. Pretty Print" << endl;
         cout << "5. Main Menu" << endl;
-        
+
         getline(cin, inputLine);
-        secondChoice = stoi(inputLine);
-        
-        switch (secondChoice)
+        if (isdigit(inputLine[0]))
         {
-            case 1:
-                tree.printPostOrder();
-                break;
-            case 2:
-                tree.printPreOrder();
-                break;
-            case 3:
-                tree.printInOrder();
-                break;
-            case 4:
-                tree.prettyPrint();
-                break;
-            case 5:
-                break;
+            secondChoice = stoi(inputLine);
+
+            switch (secondChoice)
+            {
+                case 1:
+                    tree.printPostOrder();
+                    break;
+                case 2:
+                    tree.printPreOrder();
+                    break;
+                case 3:
+                    tree.printInOrder();
+                    break;
+                case 4:
+                    tree.prettyPrint();
+                    break;
+                case 5:
+                    break;
+            }
+        }
+        else
+        {
+            cout << "Please input an integer" << endl;
         }
     }
 }
@@ -203,13 +262,13 @@ void loadFileData(BinaryTree& tree, const char* filename)
     int value;
     ifstream myfile;
     myfile.open(filename);
-    
+
     if (!myfile.is_open())
     {
         cout << "File not found." << endl;
         return;
     }
-    
+
     while (!myfile.eof())
     {
         myfile >> value;
@@ -220,49 +279,56 @@ void loadFileData(BinaryTree& tree, const char* filename)
 
 int main(int argc, const char* argv[]) {
     int mainChoice = 0;
-    
+
     BinaryTree tree; // creates the tree
-    
+
     if (argc != 2) // checks for the correct number of arguments
     {
         cout << "Invalid number of arguments." << endl;
         return 1;
     }
-    
+
     loadFileData(tree, argv[1]); // loads the tree with the information from the file
-    
+
     while (mainChoice != 7) // contains calls to each separate menu
     {
         mainMenu();
-        
+
         string inputLine;
         getline(cin, inputLine);
-        mainChoice = stoi(inputLine);
-        
-        switch (mainChoice)
+        if (isdigit(inputLine[0]))
         {
-            case 1:
-                functionsMenu(tree);
-                break;
-            case 2:
-                informationMenu(tree);
-                break;
-            case 3:
-                operationsMenu(tree);
-                break;
-            case 4:
-                printingMenu(tree);
-                break;
-            case 5:
-                tree.deleteAll();
-                loadFileData(tree, argv[1]);
-                break;
-            case 6:
-                tree.deleteAll();
-                break;
-            case 7:
-                cout << "Goodbye!" << endl;
-                break;
+            mainChoice = stoi(inputLine);
+
+            switch (mainChoice)
+            {
+                case 1:
+                    functionsMenu(tree);
+                    break;
+                case 2:
+                    informationMenu(tree);
+                    break;
+                case 3:
+                    operationsMenu(tree);
+                    break;
+                case 4:
+                    printingMenu(tree);
+                    break;
+                case 5:
+                    tree.deleteAll();
+                    loadFileData(tree, argv[1]);
+                    break;
+                case 6:
+                    tree.deleteAll();
+                    break;
+                case 7:
+                    cout << "Goodbye!" << endl;
+                    break;
+            }
+        }
+        else
+        {
+            cout << "Please input an integer" << endl;
         }
     }
     return 0;
